Added introductory Asymptote slide presentation (intro.pdf).
Added keepaux option to keep intermediate LaTeX aux files.

Added example filegraph.asy of graphing columns of data from a file.

+++ picture.cc
@@ -246,7 +246,7 @@ bool picture::texprocess(const string& texname, const string& outname,
       
     if(!getSetting<bool>("keep")) { // Delete temporary files.
       unlink(texname.c_str());
-      unlink(auxname(prefix,"aux").c_str());
+      if(!getSetting<bool>("keepaux")) unlink(auxname(prefix,"aux").c_str());
       unlink(auxname(prefix,"log").c_str());
       unlink(auxname(prefix,"out").c_str());
     }
